+++ bundle/tests/include_default/resource.yml
