+++ bundle/tests/include_multiple/databricks.yml
@@ -1,0 +1,5 @@
+bundle:
+  name: include_default
+
+include:
+  - "*/*.yml"
