+++ bundle/tests/include_multiple/my_second_job/resource.yml
@@ -1,0 +1,4 @@
+resources:
+  jobs:
+    my_second_job:
+      id: 2
