--- bundle/tests/include_override_test.go
@@ -1,12 +1,0 @@
-package config_tests
-
-import (
-	"testing"
-
-	"github.com/stretchr/testify/assert"
-)
-
-func TestIncludeOverride(t *testing.T) {
-	b := load(t, "./include_override")
-	assert.Empty(t, b.Config.Resources.Jobs)
-}
